# docs/swagger.yaml
openapi: 3.0.3
security:
  - BearerAuth: []
info:
  title: AI-Persona Backend API
  version: 1.0.0
  description: |
    # AI-Persona SaaS Backend API - Core Authentication & User Management

    ## **Overview**
    This API provides enterprise-grade authentication and user management for the AI-Persona SaaS platform. The core module is designed for production environments with advanced security features and compliance.

    ## **Core Features**
    - **Authentication**: Local (email/password) and Google OAuth authentication
    - **Account Lifecycle**: PENDING_VERIFY → ACTIVE → DEACTIVATED → PENDING_DELETION
    - **Email Verification**: 24-hour TTL verification tokens with resend capability
    - **Password Reset**: Secure token-based password reset with 1-hour TTL
    - **Session Management**: Device tracking, session revocation, and token rotation
    - **Account Lockout**: Automatic lockout after failed login attempts (5 attempts → 15 min)
    - **GDPR Compliance**: Account deletion with 30-day grace period
    - **Webhook Integration**: External service integration for persona trait updates

    ## **Security Architecture**
    - **JWT Security**: RSA-256 signing with key rotation and JWKS endpoint
    - **Password Security**: Strength validation + HIBP breach detection
    - **Account Protection**: Rate limiting, lockout, session management
    - **Input Validation**: Comprehensive sanitization and validation
    - **Audit Trail**: Complete event logging with IP, user agent, trace IDs

...
    ```
...

...

...

...

...
servers:
...
paths:
...
  /api/workspaces/{id}:
    get:
      tags:
        - Workspace Management
      summary: Get workspace details
      description: Get workspace information for the current user's workspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Workspace ID
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
                properties:
                  data:
                    type: object
                    properties:
                      workspace:
                        $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Workspace Management
      summary: Update workspace
      description: Update workspace settings (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Workspace ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "My Workspace"
                timezone:
                  type: string
                  example: "America/New_York"
                locale:
                  type: string
                  example: "en-US"
      responses:
        "200":
          description: Workspace updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
                properties:
                  data:
                    type: object
                    properties:
                      workspace:
                        $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          example: "workspace_id_here"
        name:
          type: string
          example: "My Workspace"
        domain:
          type: string
          example: "example.com"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Bad request"
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: "email"
                    message:
                      type: string
                      example: "Invalid email format"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Authentication required"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Insufficient permissions"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Resource not found"
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Validation failed"
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: "email"
                    message:
                      type: string
                      example: "Invalid email format"
...